requirements.txt
Pillow>=10.0.0
requests>=2.31.0
torch>=2.0.1
torchvision>=0.15.0 